+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Latest
 * Create separate Region (state) model from Location.
 * Create separate SubRegion (county) model from Location.
 * Add Area model for defining groups of Locations.
+* Mirror Country, Region, District and Area on Checklist and Observation
+  to speed up queries.
 
 1.0.3 (2025-03-20)
 ------------------